Changes to reports and added test

+++ src/main/java/com/santosh/ReportByHour.java
@@ -4,7 +4,7 @@ import javax.xml.bind.annotation.XmlRootElement;
 
 @XmlRootElement
 public class ReportByHour {
-    private int no_of_hours;
+    private long no_of_hours;
     private int payments;
 
     public int getPayments() {
@@ -15,11 +15,11 @@ public class ReportByHour {
         this.payments = payments;
     }
 
-    public int getNo_of_hours() {
+    public long getNo_of_hours() {
         return no_of_hours;
     }
 
-    public void setNo_of_hours(int no_of_hours) {
+    public void setNo_of_hours(long no_of_hours) {
         this.no_of_hours = no_of_hours;
     }
 }
